impede que um brinde seja deletado caso existe um ganhador

+++ app/Http/Controllers/BrindeController.php
@@ -102,12 +102,20 @@ class BrindeController extends Controller
     public function destroy(string $uid)
     {
         $brinde = Brinde::find($uid);
-        $brinde->deleted_by = Auth::user()->name;
-        $this->removeImage($brinde->imagem);
-        $brinde->imagem = null;
-        $brinde->save();
-        $brinde->delete();
-        return redirect()->to('/brindes')->with('message', 'Brinde removido com sucesso.');
+        if( empty( $brinde->funcionario_uid ) ) {
+            $brinde->deleted_by = Auth::user()->name;
+            $this->removeImage($brinde->imagem);
+            $brinde->imagem = null;
+            $brinde->save();
+            $brinde->delete();
+            $action = 'message';
+            $message = 'Brinde removido com sucesso.';
+        } else {
+            $action = 'error';
+            $message = 'Registro não pode ser deletado pois já existe um ganhador vinculado.';
+        }
+
+        return redirect()->to('/brindes')->with($action, $message);
     }
